pyproject.toml
[project]
name = "my-django-project"
version = "0.1.0"
description = "service for counting user's expenses, version: white rabbit"
authors = [
    {name = "Banakh Andrei Maksimovich", email = "banakh.andrei@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django==4.1.7",
    "djangorestframework==3.14.0", 
    "django-filter==23.2",
    "python-decouple>=3.8,<4.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "djangorestframework-simplejwt==5.2.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
